requirements.txt
pandas
plotly
python-dateutil
nltk